+++ src/components/mypage/EditProfileModal.tsx
@@ -1,6 +1,5 @@
-// EditProfileModal.tsx
 import { useEffect, useState } from "react";
-import ProfileImageUpload from "./ProfileImageUpload";
+import ProfileImageUpload from "../ProfileImageUpload";
 
 interface EditProfileModalProps {
   isOpen: boolean;
